Convert Home to functional component with hooks. Add toolbar to header. Refactor playbook card.
