+++ src/routes/Home.js
@@ -1,12 +1,12 @@
-import React, { Component } from 'react';
-import { withRouter } from 'react-router-dom';
-import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
+import React, { useState, useEffect } from 'react';
+import { useDispatch, useSelector as reduxSelector } from 'react-redux';
 import { StackItem, Stack } from '@patternfly/react-core';
+import GlobalSkeleton from '../skeletons/GlobalSkeleton';
 
 import * as actions from '../actions';
+import { downloadPlaybook } from '../api';
 
-import { Main, PageHeader, PageHeaderTitle, Wizard } from '@redhat-cloud-services/frontend-components';
+import { Main, PageHeader, PageHeaderTitle, PrimaryToolbar, Wizard } from '@redhat-cloud-services/frontend-components';
 import RemediationTable from '../components/RemediationTable';
 import TestButtons from '../components/TestButtons';
 
@@ -21,36 +21,87 @@ import './Home.scss';
 import { PermissionContext } from '../App';
 import DeniedState from '../components/DeniedState';
 import NoReceptorBanner from '../components/Alerts/NoReceptorBanner';
+import { useFilter, usePagination, useSelector, useSorter } from '../hooks/table';
+import ConfirmationDialog from '../components/ConfirmationDialog';
+import { keyBy } from 'lodash/keyBy';
 
-const ConnectedRemediationTable = connect(({ remediations }) => ({ ...remediations }))(RemediationTable);
+function skeleton () {
+    return (
+        <GlobalSkeleton></GlobalSkeleton>
+    );
+}
 
-class Home extends Component {
+function downloadAll (selectedIds, data) {
+    const byId = keyBy(data, r => r.id);
+    selectedIds.reduce((result, id) => {
+        const remediation = byId[id];
 
-    constructor (props, ctx) {
-        super(props, ctx);
-        this.state = {
-            isModalOpen: false,
-            noReceptorBannerVisible: (localStorage.getItem('remediations:receptorBannerStatus') !== 'dismissed'),
-            selected: []
-        };
-    };
+        if (remediation && remediation.issue_count === 0) {
+            return result;
+        }
 
-    handleNoReceptorToggle = () => {
-        this.setState({
-            noReceptorBannerVisible: false
-        });
+        return result.then(() => downloadPlaybook(id));
+    }, Promise.resolve());
+}
+
+const SORTING_ITERATEES = [ null, 'name', 'system_count', 'issue_count', 'updated_at' ];
+
+function Home () {
+
+    const [ isModalOpen, setIsModalOpen ] = useState(false);
+    const [ noReceptorBannerVisible, setNoReceptorBannerVisible ] = useState(
+        localStorage.getItem('remediations:receptorBannerStatus') !== 'dismissed');
+    const sorter = useSorter(4, 'desc');
+    const filter = useFilter();
+    const selector = useSelector();
+    const pagination = usePagination();
+    const [ filterText, setFilterText ] = useState('');
+    const [ dialogOpen, setDialogOpen ] = useState(false);
+    const [ showArchived, setShowArchived ] = useState(false);
+    const [ shouldUpdateGrid, setShouldUpdateGrid ] = useState(false);
+    const remediations = reduxSelector(state => state.remediations);
+    const dispatch = useDispatch();
+    const loadRemediations = (...args) => dispatch(actions.loadRemediations(...args));
+    const deleteRemediation = (id) => dispatch(actions.deleteRemediation(id));
+
+    function load () {
+        const column = SORTING_ITERATEES[sorter.sortBy];
+        loadRemediations(column, sorter.sortDir, filter.value, pagination.pageSize, pagination.offset);
+    }
+
+    useEffect(load, []);
+
+    useEffect(() => {
+        if (remediations.status === 'fulfilled') {
+            setShouldUpdateGrid(true);
+        }
+    }, [ sorter.sortBy, sorter.sortDir, filter.value, pagination.pageSize, pagination.pageDebounced ]);
+
+    if (remediations.status !== 'fulfilled') {
+        return skeleton();
+    }
+
+    filter.onChange(pagination.reset);
+    sorter.onChange(pagination.reset);
+
+    const selectedIds = selector.getSelectedIds();
+
+    const handleNoReceptorToggle = () => {
+        setNoReceptorBannerVisible(false);
         localStorage.setItem('remediations:receptorBannerStatus', 'dismissed');
     };
 
-    openModal = () => this.setState({ isModalOpen: true });
+    const sendNotification = (data) => {
+        dispatch(addNotification(data));
+    };
 
-    onClose = submitted => {
-        this.setState({
-            isModalOpen: false
-        });
+    // const openModal = () => setIsModalOpen(true);
+
+    const onClose = (submitted) => {
+        setIsModalOpen(false);
 
         if (submitted) {
-            this.sendNotification({
+            sendNotification({
                 variant: 'success',
                 title: 'Wizard completed',
                 description: 'Congratulations! You successfully clicked through the temporary wizard placeholder!'
@@ -58,76 +109,113 @@ class Home extends Component {
         }
     };
 
-    sendNotification = data => {
-        const { addNotification } = this.props;
-        addNotification(data);
-    }
-
-    onRemediationCreated = result => {
-        const { loadRemediations } = this.props;
-        this.sendNotification(result.getNotification());
-        loadRemediations();
+    const onRemediationCreated = (result) => {
+        sendNotification(result.getNotification());
+        dispatch(actions.loadRemediations());
     };
 
-    onSelect = selected => this.setState({ selected });
+    // Wizard Content
+    const ModalStepContent = [
+        <PlanName key='PlanName'/>,
+        <PlanSystems key='PlanSystems'/>
+    ];
 
-    render() {
-
-        const { isModalOpen } = this.state;
-        const { loadRemediations, deleteRemediation } = this.props;
-
-        // Wizard Content
-        const ModalStepContent = [
-            <PlanName key='PlanName'/>,
-            <PlanSystems key='PlanSystems'/>
-        ];
-
-        return (
-            <PermissionContext.Consumer>
-                { value =>
-                    value.permissions.read === false
-                        ? <DeniedState/>
-                        : <React.Fragment>
-                            <PageHeader>
-                                <PageHeaderTitle title='Remediations'/>
-                                <TestButtons onRemediationCreated={ this.onRemediationCreated } />
-                            </PageHeader>
-                            <Main>
-                                <Stack hasGutter>
-                                    { value.hasSmartManagement && !value.isReceptorConfigured && this.state.noReceptorBannerVisible &&
-                                        <StackItem>
-                                            <NoReceptorBanner onClose={ () => this.handleNoReceptorToggle() }/>
-                                        </StackItem>
+    return (
+        <PermissionContext.Consumer>
+            { permission =>
+                permission.permissions.read === false
+                    ? <DeniedState/>
+                    : <React.Fragment>
+                        <PageHeader>
+                            <PageHeaderTitle title='Remediations'/>
+                            <TestButtons onRemediationCreated={ onRemediationCreated } />
+                        </PageHeader>
+                        <PrimaryToolbar
+                            filterConfig={ {
+                                items: [{
+                                    label: 'Search playbooks',
+                                    type: 'text',
+                                    filterValues: {
+                                        id: 'filter-by-string',
+                                        key: 'filter-by-string',
+                                        placeholder: 'Search playbooks',
+                                        value: filterText,
+                                        onChange: (_e, value) => {
+                                            setFilterText(value);
+                                        },
+                                        onSubmit: (_e, value) => {
+                                            filter.setValue(value);
+                                        }
                                     }
+                                }]
+                            } }
+                            bulkSelect={ { items: [{ title: 'Select all',
+                                onClick: (e) => selector.props.onSelect(e, true, -1)
+                            }],
+                            checked: selectedIds.length && remediations.value.meta.total > selectedIds.length ? null : selectedIds.length,
+                            count: selectedIds.length,
+                            onSelect: (isSelected, e) => selector.props.onSelect(e, isSelected, -1) } }
+                            actionsConfig={ { actions: [
+                                { label: 'Download playbooks',
+                                    props: { variant: 'secondary', isDisabled: !selectedIds.length },
+                                    onClick: () => downloadAll(selectedIds, remediations.value.data)
+                                },
+                                { label: 'Delete playbooks',
+                                    props: { isDisabled: !permission.permissions.write || !selectedIds.length },
+                                    onClick: () => setDialogOpen(true)
+                                },
+                                { label: showArchived ? 'Hide archived playbooks' : 'Show archived playbooks',
+                                    onClick: showArchived ? () => setShowArchived(false) : () => setShowArchived(true)
+                                }]} }
+                            pagination={ { ...pagination.props, itemCount: remediations.value.meta.total } }
+                        />
+                        <Main>
+                            <Stack hasGutter>
+                                { permission.hasSmartManagement && !permission.isReceptorConfigured && noReceptorBannerVisible &&
                                     <StackItem>
-                                        <ConnectedRemediationTable loadRemediations={ loadRemediations } deleteRemediation={ deleteRemediation }/>
+                                        <NoReceptorBanner onClose={ () => handleNoReceptorToggle() }/>
                                     </StackItem>
-                                </Stack>
-                            </Main>
+                                }
+                                { dialogOpen &&
+                                    <ConfirmationDialog
+                                        text={ `You will not be able to recover ${selectedIds.length > 1
+                                            ? 'these remediations' : 'this remediation'}` }
+                                        onClose={ async (del) => {
+                                            setDialogOpen(false);
+                                            if (del) {
+                                                await Promise.all(selectedIds.map(r => deleteRemediation(r)));
+                                                loadRemediations();
+                                                selector.reset();
+                                            }
+                                        } } />
+                                }
+                                <StackItem>
+                                    <RemediationTable
+                                        remediations={ remediations }
+                                        loadRemediations={ loadRemediations }
+                                        sorter={ sorter }
+                                        filter={ filter }
+                                        selector={ selector }
+                                        pagination={ pagination }
+                                        shouldUpdateGrid={ shouldUpdateGrid }
+                                        setShouldUpdateGrid={ setShouldUpdateGrid }
+                                    />
+                                </StackItem>
+                            </Stack>
+                        </Main>
 
-                            <Wizard
-                                isLarge
-                                title="Create Plan"
-                                className='ins-c-plan-modal'
-                                onClose = { this.onClose }
-                                isOpen= { isModalOpen }
-                                content = { ModalStepContent }
-                            />
-                        </React.Fragment>
-                }
-            </PermissionContext.Consumer>
-        );
-    }
+                        <Wizard
+                            isLarge
+                            title="Create Plan"
+                            className='ins-c-plan-modal'
+                            onClose = { onClose }
+                            isOpen= { isModalOpen }
+                            content = { ModalStepContent }
+                        />
+                    </React.Fragment>
+            }
+        </PermissionContext.Consumer>
+    );
 }
 
-Home.propTypes = {
-    loadRemediations: PropTypes.func,
-    deleteRemediation: PropTypes.func,
-    addNotification: PropTypes.func
-};
-
-export default withRouter(connect(null, (dispatch) => ({
-    loadRemediations: (...args)  => dispatch(actions.loadRemediations(...args)),
-    deleteRemediation: (id) => dispatch(actions.deleteRemediation(id)),
-    addNotification: (data) => dispatch(addNotification(data))
-}))(Home));
+export default Home;
